AP-5888: possibly squash into prev
