+++ app/helpers/application_helper.rb
@@ -1,12 +1,4 @@
 module ApplicationHelper
-  def home_path
-    request.path_info.include?("providers") ? your_applications_default_tab_path : "#"
-  end
-
-  def your_applications_default_tab_path
-    submitted_providers_legal_aid_applications_path
-  end
-
   def html_title
     default = t("shared.page-title.suffix")
     return default unless content_for?(:head_title) || content_for?(:page_title)
