+++ app/helpers/your_applications_helper.rb
@@ -1,5 +1,7 @@
 module YourApplicationsHelper
-  extend ActiveSupport::Concern
+  def home_path
+    request.path_info.include?("providers") ? your_applications_default_tab_path : "#"
+  end
 
   def your_applications_default_tab_path
     submitted_providers_legal_aid_applications_path
